mglass: try to fix the bug in magnifiers::nearestNeighbor (2).
